+++ Robot_Adapter/Convert/ToRobot/Loads/LoadcaseNature.cs
@@ -20,53 +20,49 @@
  * along with this code. If not, see <https://www.gnu.org/licenses/lgpl-3.0.html>.      
  */
 
+using BH.oM.Structure.Loads;
 using RobotOM;
-using BH.oM.Structure.Constraints;
-using System.Collections.Generic;
-using System;
-using BH.oM.Structure.SurfaceProperties;
 
-namespace BH.Engine.Robot
+namespace BH.Adapter.Robot
 {
-    public class LoadingPanelPropertyComparer : IEqualityComparer<LoadingPanelProperty>
+    public static partial class Convert
     {
         /***************************************************/
         /****           Public Methods                  ****/
         /***************************************************/
 
-        public bool Equals(LoadingPanelProperty property1, LoadingPanelProperty property2)
+        public static IRobotCaseNature ToRobotLoadcaseNature(Loadcase lCase, out int subNature)
         {
-            //Check whether the compared objects reference the same data.
-            if (Object.ReferenceEquals(property1, property2))
-                return true;
-
-            //Check whether any of the compared objects is null.
-            if (Object.ReferenceEquals(property1, null) || Object.ReferenceEquals(property2, null))
-                return false;
-
-            //Check if the GUIDs are the same
-            if (property1.BHoM_Guid == property2.BHoM_Guid)
-                return true;
-
-            if (property1.Name == property2.Name &&
-                 property1.Material.Name == property2.Material.Name &&
-                 property1.LoadApplication == property2.LoadApplication)
-                return true;
-            return false;
+            subNature = -1;
+            switch (lCase.Nature)
+            {
+                case LoadNature.Dead:
+                case LoadNature.SuperDead:
+                    subNature = 1;
+                    return IRobotCaseNature.I_CN_PERMANENT;
+                case LoadNature.Live:
+                    subNature = 4;
+                    return IRobotCaseNature.I_CN_EXPLOATATION;
+                case LoadNature.Accidental:
+                    subNature = 15;
+                    return IRobotCaseNature.I_CN_ACCIDENTAL;
+                case LoadNature.Snow:
+                    subNature = 10;
+                    return IRobotCaseNature.I_CN_SNOW;
+                case LoadNature.Wind:
+                    subNature = 13;
+                    return IRobotCaseNature.I_CN_WIND;
+                case LoadNature.Temperature:
+                    subNature = 14;
+                    return IRobotCaseNature.I_CN_TEMPERATURE;
+                case LoadNature.Seismic:
+                    return IRobotCaseNature.I_CN_SEISMIC;
+                default:
+                    return IRobotCaseNature.I_CN_PERMANENT;
+            }
         }
 
         /***************************************************/
-
-        public int GetHashCode(LoadingPanelProperty obj)
-        {
-            //Check whether the object is null
-            if (Object.ReferenceEquals(obj, null)) return 0;
-
-            return obj.Name == null ? 0 : obj.Name.GetHashCode();
-        }
-
-        /***************************************************/
-       
     }
 }
