+++ Robot_Adapter/Convert/ToRobot/Loads/PointAcceleration.cs
@@ -20,53 +20,34 @@
  * along with this code. If not, see <https://www.gnu.org/licenses/lgpl-3.0.html>.      
  */
 
+using BH.Engine.Geometry;
+using BH.oM.Structure.Loads;
 using RobotOM;
-using BH.oM.Structure.Constraints;
-using System.Collections.Generic;
-using System;
-using BH.oM.Structure.SurfaceProperties;
 
-namespace BH.Engine.Robot
+namespace BH.Adapter.Robot
 {
-    public class LoadingPanelPropertyComparer : IEqualityComparer<LoadingPanelProperty>
+    public static partial class Convert
     {
         /***************************************************/
         /****           Public Methods                  ****/
         /***************************************************/
-
-        public bool Equals(LoadingPanelProperty property1, LoadingPanelProperty property2)
+         
+        public static void ToRobot(this PointAcceleration load, RobotSimpleCase sCase, RobotGroupServer rGroupServer)
         {
-            //Check whether the compared objects reference the same data.
-            if (Object.ReferenceEquals(property1, property2))
-                return true;
+            if (load.TranslationalAcceleration.Length() == 0)
+            {
+                Engine.Reflection.Compute.RecordError("Zero point accelerations are not pushed to Robot");
+                return;
+            }
+            IRobotLoadRecord loadRecord = sCase.Records.Create(IRobotLoadRecordType.I_LRT_NODE_ACCELERATION);
+            loadRecord.Objects.FromText(load.CreateIdListOrGroupName(rGroupServer));
+            loadRecord.SetValue((short)IRobotNodeAccelerationRecordValues.I_NACRV_UX, load.TranslationalAcceleration.X);
+            loadRecord.SetValue((short)IRobotNodeAccelerationRecordValues.I_NACRV_UY, load.TranslationalAcceleration.Y);
+            loadRecord.SetValue((short)IRobotNodeAccelerationRecordValues.I_NACRV_UZ, load.TranslationalAcceleration.Z);
 
-            //Check whether any of the compared objects is null.
-            if (Object.ReferenceEquals(property1, null) || Object.ReferenceEquals(property2, null))
-                return false;
-
-            //Check if the GUIDs are the same
-            if (property1.BHoM_Guid == property2.BHoM_Guid)
-                return true;
-
-            if (property1.Name == property2.Name &&
-                 property1.Material.Name == property2.Material.Name &&
-                 property1.LoadApplication == property2.LoadApplication)
-                return true;
-            return false;
         }
 
         /***************************************************/
-
-        public int GetHashCode(LoadingPanelProperty obj)
-        {
-            //Check whether the object is null
-            if (Object.ReferenceEquals(obj, null)) return 0;
-
-            return obj.Name == null ? 0 : obj.Name.GetHashCode();
-        }
-
-        /***************************************************/
-       
     }
 }
