+++ Robot_Adapter/Robot_Adapter.csproj
@@ -139,7 +139,20 @@
     <Compile Include="Convert\FromRobot\Geometry\Polyline.cs" />
     <Compile Include="Convert\FromRobot\Geometry\Point.cs" />
     <Compile Include="Convert\FromRobot\Geometry\Polycurve.cs" />
-    <Compile Include="Convert\ToRobot\Loads\Load.cs" />
+    <Compile Include="Convert\ToRobot\Loads\GeometricalLineLoad.cs" />
+    <Compile Include="Convert\ToRobot\Loads\ContourLoad.cs" />
+    <Compile Include="Convert\ToRobot\Loads\BarTemperatureLoad.cs" />
+    <Compile Include="Convert\ToRobot\Loads\BarVaryingDistributedLoad.cs" />
+    <Compile Include="Convert\ToRobot\Loads\ILoad.cs" />
+    <Compile Include="Convert\ToRobot\Loads\PointAcceleration.cs" />
+    <Compile Include="Convert\ToRobot\Loads\PointVelocity.cs" />
+    <Compile Include="Convert\ToRobot\Loads\PointDisplacement.cs" />
+    <Compile Include="Convert\ToRobot\Loads\BarPointLoad.cs" />
+    <Compile Include="Convert\ToRobot\Loads\AreaUniformlyDistributedLoad.cs" />
+    <Compile Include="Convert\ToRobot\Loads\GravityLoad.cs" />
+    <Compile Include="Convert\ToRobot\Loads\BarUniformlyDistributedLoad.cs" />
+    <Compile Include="Convert\ToRobot\Loads\PointLoad.cs" />
+    <Compile Include="Convert\ToRobot\Loads\LoadcaseNature.cs" />
     <Compile Include="Convert\FromRobot\Properties\LinearRelease.cs" />
     <Compile Include="Convert\FromRobot\Properties\BarRelease.cs" />
     <Compile Include="Convert\FromRobot\Properties\Material.cs" />
